requirements.txt
django==4.2.2
python-dotenv==1.1.1
psycopg2-binary==2.9.10
django-filter==25.1
